go back to slots approach
